MAT-7060 moved DTO to java-models
