--- src/main/java/gov/cms/madie/cqllibraryservice/dto/LibraryUsage.java
@@ -1,14 +1,0 @@
-package gov.cms.madie.cqllibraryservice.dto;
-
-import gov.cms.madie.models.common.Version;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class LibraryUsage {
-  private String name;
-  // TODO: getting version in string would be beneficial
-  private Version version;
-  private String owner;
-}
