+++ src/test/java/gov/cms/madie/cqllibraryservice/services/MeasureServiceClientTest.java
@@ -1,8 +1,8 @@
 package gov.cms.madie.cqllibraryservice.services;
 
 import gov.cms.madie.cqllibraryservice.config.EnvironmentConfig;
-import gov.cms.madie.cqllibraryservice.dto.LibraryUsage;
 import gov.cms.madie.cqllibraryservice.exceptions.MeasureServiceException;
+import gov.cms.madie.models.dto.LibraryUsage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
